+++ backend/src/main/java/com/cruru/applyform/service/ApplyFormService.java
@@ -11,7 +11,7 @@ import com.cruru.applyform.controller.dto.ApplyFormSubmitRequest;
 import com.cruru.applyform.domain.ApplyForm;
 import com.cruru.applyform.domain.repository.ApplyFormRepository;
 import com.cruru.applyform.exception.ApplyFormNotFoundException;
-import com.cruru.applyform.exception.PersonalDataProcessingException;
+import com.cruru.applyform.exception.badrequest.PersonalDataProcessingException;
 import com.cruru.dashboard.domain.Dashboard;
 import com.cruru.process.domain.Process;
 import com.cruru.process.domain.repository.ProcessRepository;
